fix timer button size change bug
With larger font sizes using the system font, the extra spacing becomes big enough to increase button height. This behavior is unexpectedly seen only when changing text inside the target on_button_pressed signal function as using a separate signal function to change the button text doesn't change the size. In this particular application, The button height inconsistently changes from 55px to 63px when button.text is changed on press (despite the same font size of 40px). This is fixed by creating a FontVariation  theme to set both spacing_top and spacing_bottom to -4 to remove those 8px of extra space that is vertically expanding the button. There is some inconsistent or unclear behavior updating button.text from within the on_button_pressed signal function and outside of that function that results in this issue.